api/requirements.txt
drf-yasg==1.20.0
Django==3.1.3
django-redis==4.12.1
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6